requirements.txt
asgiref==3.8.1
dj-database-url==3.0.1
Django==5.2.3
django-cors-headers==4.7.0
django-stubs==5.2.1
django-stubs-ext==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
sqlparse==0.5.3
types-PyYAML==6.0.12.20250516
typing_extensions==4.14.0
tzdata==2025.2
whitenoise==6.9.0
